pyproject.toml
[tool.poetry]
name = "mafic"
version = "0.1.0"
description = "A properly typehinted lavalink client for discord.py, nextcord, disnake and py-cord."
authors = ["ooliver1 <oliverwilkes2006@icloud.com>"]
license = "MIT"
readme = "README.md"

packages = [
    { include = "mafic" },
]

[tool.poetry.dependencies]
python = "^3.8"
aiohttp = "^3.6.0"
nextcord = {version = "^2.0.0", optional = true}

[tool.poetry.group.dev.dependencies]
pyright = "^1.1.270"
black = "^22.8.0"
isort = "^5.10.1"
taskipy = "^1.10.3"
flake8 = "^5.0.4"
pre-commit = "^2.20.0"
python-dotenv = "^0.21.0"
slotscheck = "^0.15.0"
Flake8-pyproject = "^1.1.0.post0"
typing-extensions = "^4.3.0"

[tool.poetry.group.lint.dependencies]
nextcord = "^2.0.0"
disnake = "^2.0.0"
py-cord = "^2.0.0"
"discord.py" = "^2.0.0"
orjson = "^3.8.0"

[tool.poetry.extras]
nextcord = ["nextcord@^2.0.0"]
disnake = ["disnake@^2.0.0"]
py-cord = ["py-cord@^2.0.0"]
"discord.py" = ["discord.py@^2.0.0"]

[tool.taskipy.tasks]
pre-commit = "pre-commit install --install-hooks"
slotscheck = "slotscheck --verbose -m mafic"
lint = "pre-commit run --all-files"
black = "task lint black"
isort = "task lint isort"
flake8 = "task lint flake8"
pyright = "dotenv -f task.env run -- pyright"

[tool.black]
line-length = 88
target-version = ["py38", "py39", "py310"]

[tool.isort]
profile = "black"
py_version = 38
line_length = 88
combine_as_imports = true
filter_files = true

[tool.slotscheck]
strict-imports = true
require-superclass = true
require-subclass = false

[tool.pyright]
typeCheckingMode = "strict"
pythonVersion = "3.8"

[tool.flake8]
max-line-length = 88
per-file-ignores = [
    '__init__.py:F401,F403',
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
